mv Kore.Attribute.Symbol.Symbol Kore.Attribute.Symbol
